Updated block rotation to be generic
